+++ nexus-security/src/main/java/com/github/nexus/ssl/trust/WhiteListTrustManager.java
@@ -1,4 +1,4 @@
-package com.github.nexus.ssl;
+package com.github.nexus.ssl.trust;
 
 import com.github.nexus.ssl.util.CertificateUtil;
 
@@ -7,9 +7,9 @@ import java.io.IOException;
 import java.security.cert.CertificateException;
 import java.security.cert.X509Certificate;
 
-public class TrustOnFirstUseManager extends ExtendedTrustManager {
+public class WhiteListTrustManager extends ExtendedTrustManager {
 
-    public TrustOnFirstUseManager(File knownHosts) throws IOException {
+    public WhiteListTrustManager(File knownHosts) throws IOException {
         super(knownHosts);
     }
 
@@ -30,15 +30,10 @@ public class TrustOnFirstUseManager extends ExtendedTrustManager {
 
     private void checkTrusted(X509Certificate[] x509Certificates) throws CertificateException{
         final X509Certificate certificate = x509Certificates[0];
-        final String fingerPrint = CertificateUtil.generateFingerprint(certificate);
-
-        if (!certificateExistsInKnownHosts(fingerPrint)){
-            try {
-                addServerToKnownHostsList(fingerPrint);
-            }
-            catch (IOException ex){
-                throw new CertificateException("Failed to save address and certificate fingerprint to whitelist");
-            }
+        final String thumbPrint = CertificateUtil.create().thumbPrint(certificate);
+        if (!certificateExistsInKnownHosts(thumbPrint)) {
+            throw new CertificateException("Connections not allowed");
         }
     }
+
 }
